# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AddToGroupRequest:
    properties:
      group_id:
        example: 1
        type: integer
    required:
    - group_id
    type: object
  handlers.AddToOrganizationRequest:
    properties:
      organization_id:
        example: 1
        type: integer
    required:
    - organization_id
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  models.Group:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Administrators
        type: string
      organizations:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - name
    type: object
  models.Organization:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: Acme Corp
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - name
    type: object
  models.User:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      email:
        example: john@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      organizations:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    required:
    - email
    - name
    type: object
  models.UserCreate:
    properties:
      active:
        example: true
        type: boolean
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.UserResponse:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      email:
        example: john@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      organizations:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  models.UserUpdate:
    properties:
      active:
        example: false
        type: boolean
      email:
        example: john.updated@example.com
        type: string
      name:
        example: John Doe Updated
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@kitamanager.example.com
    name: API Support
  description: REST API for managing Users, Groups, and Organizations with RBAC support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: KitaManager API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /api/v1/users/{id}/groups:
    post:
      consumes:
      - application/json
      description: Add a user to a group
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddToGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add user to group
      tags:
      - users
  /api/v1/users/{id}/groups/{gid}:
    delete:
      consumes:
      - application/json
      description: Remove a user from a group
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group ID
        in: path
        name: gid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove user from group
      tags:
      - users
  /api/v1/users/{id}/organizations:
    post:
      consumes:
      - application/json
      description: Add a user to an organization
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddToOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add user to organization
      tags:
      - users
  /api/v1/users/{id}/organizations/{oid}:
    delete:
      consumes:
      - application/json
      description: Remove a user from an organization
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization ID
        in: path
        name: oid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove user from organization
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"